fix(filters): filter layout revamped, global context with filter type
