--- src/app/(tabs)/(cards)/card-list-index.tsx
@@ -1,26 +1,0 @@
-import { StyleSheet, View } from 'react-native';
-import { useTheme } from '../../../theme/useTheme';
-import CardsList from '../../../features/cards/screens/CardsList';
-
-export default function CardListIndex() {
-  const theme = useTheme();
-
-  return (
-    <View style={[styles.container, { backgroundColor: theme.colors.background.primary }]}>
-      <View style={[styles.content, { backgroundColor: theme.colors.background.primary }]}>
-        <CardsList isFiltersVisible={true} />
-      </View>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  content: {
-    flex: 1,
-    paddingHorizontal: 16,
-    paddingTop: 16,
-  },
-});
