--- src/app/(tabs)/(sets)/filter-set-cards-detail.tsx
@@ -1,6 +1,0 @@
-import FilterDetail from '../../../features/filters/components/FilterDetail';
-import { FilterTypeEnum } from '../../../features/filters/types';
-
-export default function FilterSetCardsDetailScreen() {
-  return <FilterDetail filterType={FilterTypeEnum.SET} />;
-}
