+++ src/app/(tabs)/index.tsx
@@ -1,5 +1,26 @@
-import { Redirect } from 'expo-router';
+import { StyleSheet, View } from 'react-native';
+import { useTheme } from '../../theme/useTheme';
+import CardsList from '../../features/cards/screens/CardsList';
 
 export default function Index() {
-  return <Redirect href="/(cards)/card-list-index" />;
+  const theme = useTheme();
+
+  return (
+    <View style={[styles.container, { backgroundColor: theme.colors.background.primary }]}>
+      <View style={[styles.content, { backgroundColor: theme.colors.background.primary }]}>
+        <CardsList isFiltersVisible={true} />
+      </View>
+    </View>
+  );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  content: {
+    flex: 1,
+    paddingHorizontal: 16,
+    paddingTop: 16,
+  },
+});
