+++ src/app/(tabs)/scan.tsx
@@ -1,5 +1,6 @@
 import CardScanner from '../../features/scan/screens/CardScanner';
 
 export default function Scan() {
-  return <CardScanner />;
+  return <></>;
+  // return <CardScanner />;
 }
